+++ src/main/java/com/cburch/logisim/std/memory/Ram.java
@@ -290,7 +290,10 @@ public class Ram extends Mem {
 
     // perform reads
     final var width = state.getAttributeValue(DATA_ATTR);
-    final var outputEnabled = separate || !state.getPortValue(RamAppearance.getOEIndex(0, attrs)).equals(Value.FALSE);
+    final var outputDisabled =
+        state.getPortValue(RamAppearance.getOEIndex(0, attrs)).equals(Value.FALSE)
+            && (!separate || RamAttributes.hasControlledOutput(attrs));
+    final var outputEnabled = !outputDisabled;
     if (outputEnabled && goodAddr && !misalignError) {
       for (var i = 0; i < dataLines; i++) {
         long val = myState.getContents().get(addr + i);
@@ -308,7 +311,6 @@ public class Ram extends Mem {
   private void propagateByteEnables(InstanceState state, long addr, boolean goodAddr, boolean errorValue) {
     final var attrs = state.getAttributeSet();
     final var myState = (RamState) getState(state);
-    final var separate = isSeparate(attrs);
     long oldMemValue = myState.getContents().get(myState.getCurrent());
     long newMemValue = oldMemValue;
     // perform writes
@@ -345,13 +347,12 @@ public class Ram extends Mem {
 
     Consumer<Value> setValue = (Value value) -> state.setPort(RamAppearance.getDataOutIndex(0, attrs), value, DELAY);
 
-    if (!separate && outputNotEnabled) {
-      /* put the bus in tri-state in case of a combined bus and no output enable */
-      setValue.accept(Value.createUnknown(dataBits));
+    if (outputNotEnabled) {
+      if (!isSeparate(attrs) || RamAttributes.hasControlledOutput(attrs)) {
+        setValue.accept(Value.createUnknown(dataBits));
+      }
       return;
     }
-    /* if the OE is not activated return */
-    if (outputNotEnabled) return;
 
     /* if the address is bogus set error value accordingly */
 
@@ -387,7 +388,7 @@ public class Ram extends Mem {
 
   public static boolean isSeparate(AttributeSet attrs) {
     Object bus = attrs.getValue(RamAttributes.ATTR_DBUS);
-    return bus == null || bus.equals(RamAttributes.BUS_SEP);
+    return RamAttributes.isSeparateDataBus(bus);
   }
 
   @Override